# app-service/swagger.yaml
basePath: /api/v1
definitions:
  controller.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  controller.TokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  controller.UserResponse:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      phone:
        type: string
      username:
        type: string
    type: object
  service.UserLoginDTO:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  service.UserPatchDTO:
    properties:
      email:
        type: string
      full_name:
        type: string
      phone:
        type: string
      username:
        type: string
    type: object
  service.UserRegisterDTO:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    required:
    - email
    - full_name
    - password
    - phone
    - username
    type: object
host: www.quible.io
info:
  contact: {}
  description: Authentication and authorization service of Quible.io
  title: Quible app-service
  version: "0.1"
paths:
  /docs:
    get:
      description: Render Swagger UI page
      parameters:
      - default: swagger
        description: UI template
        enum:
        - swagger
        - rapidoc
        - redoc
        in: query
        name: ui
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Swagger UI
      tags:
      - docs
  /docs/spec:
    get:
      description: Swagger spec in YAML format
      produces:
      - application/yaml
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Swagger spec
      tags:
      - docs
  /login:
    post:
      consumes:
      - application/json
      description: Login with user credentials to get token
      parameters:
      - description: User login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UserLoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Login
      tags:
      - user
      - public
  /user:
    get:
      description: Returns user profile associated with the token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Get user
      tags:
      - user
      - private
    patch:
      consumes:
      - application/json
      description: Updates user profile associated with the token
      parameters:
      - description: Partial user object to be used for update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UserPatchDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Update user
      tags:
      - user
      - private
    post:
      consumes:
      - application/json
      description: Register a new user.
      parameters:
      - description: User registration information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UserRegisterDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Register
      tags:
      - user
      - public
swagger: "2.0"